Routing: Fix bug of vertical parking failure.
Fix $[IDG-Apollo-4346].
When the vehicle is in front of the parking spot, and on the lane next
to the nearest lane to the parking spot, the parking action will fail
because of the following error:
Failed to extract segments from routing.
and the relative error of pnc map like:
Failed to find nearest point: x: 437868.09950172278 y: 4433040.1114324955 z: 0

Cause:
The lane segments of the routing response don't contain the lane nearest
to the parking spot because the vehicle is in front of the parking spot.
When the vehicle approach to the parking spot, no project point can be
found on the routing segments and thus CreateReferenceLine failed.

Solution:
1. Add SupplementParkingRequest function. When the routing segements
doesn't contain the lane nearest to the parking spot, add it into the
sgement list.
2. Modify the way of searching the parking spot according to the vehicle
state as well as the way of searching the nearest lane to the
vehicle(GetNearestPointFromRouting).