fix(planning): Remove "stop_distance" of CreepStage.
Refactor the process of CreepStage.
There are 2 distances for CreepStage, one is the target stop distance
where a stop fence will be built, but the vehicle will not actually stop
here, because CreepStage will finish before it and vehicle will continue
moving on; the other distance is the finish distance, where the vehicle
reaches, CreepStage will be considered finished.

Remove the unimportant parameters "stop_distance" and
"max_valid_stop_distance" because they don't have the direct relation
with the creeping behavior.

Change-Id: I00a38e679af59eee2f1144f882bc7f9583518438